+++ esp-hal/src/spi/mod.rs
@@ -12,7 +12,6 @@
 use crate::dma::DmaError;
 
 pub mod master;
-#[cfg(not(esp32))]
 pub mod slave;
 
 /// SPI errors